Split service responsibilities
