+++ Model/Service/Merchandise/RequestParamsService.php
@@ -1,41 +1,6 @@
-<?php /** @noinspection PhpDeprecationInspection */
+<?php
 
-/**
- * Copyright (c) 2020, Nosto Solutions Ltd
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- * this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors
- * may be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2020 Nosto Solutions Ltd
- * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
- *
- */
-
-namespace Nosto\Cmp\Model\Service\Recommendation;
+namespace Nosto\Cmp\Model\Service\Merchandise;
 
 use Magento\Catalog\Api\CategoryRepositoryInterface;
 use Magento\Framework\Event\ManagerInterface;
@@ -49,7 +14,7 @@ use Magento\Store\Model\Store;
 use Nosto\Cmp\Helper\Data;
 use Nosto\Cmp\Model\Facet\FacetInterface;
 use Nosto\Cmp\Model\Merchandise\MerchandiseRequestParams;
-use Nosto\Cmp\Model\Service\Session\SessionService as VisitSessionService;
+use Nosto\Cmp\Model\Service\VisitSession\SessionService as VisitSessionService;
 use Nosto\Cmp\Observer\App\Action\PostRequestAction;
 use Nosto\Cmp\Observer\App\Action\PreRequestAction;
 use Nosto\Cmp\Utils\Debug\ServerTiming;
@@ -68,16 +33,11 @@ use Nosto\Tagging\Model\Service\Product\Category\DefaultCategoryService as Categ
 use Nosto\NostoException;
 use Nosto\Request\Http\Exception\AbstractHttpException;
 use Nosto\Request\Http\Exception\HttpResponseException;
-use Nosto\Cmp\Model\Service\Recommendation\SessionService as ResultSessionService;
+use Nosto\Cmp\Model\Service\MagentoSession\SessionService as ResultSessionService;
 
-class StateAwareCategoryService
+class RequestParamsService
 {
     const NOSTO_PREVIEW_COOKIE = 'nostopreview';
-    const TIME_PROF_GRAPHQL_QUERY = 'cmp_graphql_query';
-
-    use LoggerTrait {
-        LoggerTrait::__construct as loggerTraitConstruct; // @codingStandardsIgnoreLine
-    }
 
     /**
      * @var CookieManagerInterface
@@ -105,16 +65,8 @@ class StateAwareCategoryService
     private $categoryBuilder;
 
     /**
-     * @var CategoryMerchandisingResult|null
+     * @var CategoryRepositoryInterface
      */
-    private $lastResult = null;
-
-    /**
-     * @var int
-     */
-    private $lastUsedLimit;
-
-    /** @var CategoryRepositoryInterface */
     private $categoryRepository;
 
     /**
@@ -137,19 +89,6 @@ class StateAwareCategoryService
      */
     private $resultSessionService;
 
-    /**
-     * StateAwareCategoryService constructor.
-     * @param CookieManagerInterface $cookieManager
-     * @param NostoHelperAccount $nostoHelperAccount
-     * @param NostoHelperScope $nostoHelperScope
-     * @param Registry $registry
-     * @param CategoryBuilder $categoryBuilder
-     * @param Logger $logger
-     * @param Data $nostoCmpHelper
-     * @param CategoryRepositoryInterface $categoryRepository
-     * @param ManagerInterface $eventManager
-     * @param SessionService $sessionService
-     */
     public function __construct(
         CookieManagerInterface $cookieManager,
         NostoHelperAccount $nostoHelperAccount,
@@ -163,10 +102,6 @@ class StateAwareCategoryService
         VisitSessionService $visitSessionService,
         ResultSessionService $resultSessionService
     ) {
-        $this->loggerTraitConstruct(
-            $logger
-        );
-        $this->cookieManager = $cookieManager;
         $this->cookieManager = $cookieManager;
         $this->nostoHelperAccount = $nostoHelperAccount;
         $this->nostoHelperScope = $nostoHelperScope;
@@ -180,32 +115,6 @@ class StateAwareCategoryService
     }
 
     /**
-     * @inheritDoc
-     * @throws MissingAccountException
-     * @throws MissingTokenException
-     * @throws SessionCreationException
-     */
-    public function getPersonalisationResult(
-        FacetInterface $facets,
-        $pageNumber,
-        $limit
-    ): ?CategoryMerchandisingResult {
-
-        $requestParams = $this->createRequestParams($facets, $pageNumber, $limit);
-
-        $this->lastResult = ServerTiming::getInstance()->instrument(
-            function () use ($requestParams) {
-                return $this->getMerchandiseResults($requestParams);
-            },
-            self::TIME_PROF_GRAPHQL_QUERY
-        );
-
-        $this->lastUsedLimit = $limit;
-
-        return $this->lastResult;
-    }
-
-    /**
      * @param FacetInterface $facets
      * @param $pageNumber
      * @param $limit
@@ -214,7 +123,7 @@ class StateAwareCategoryService
      * @throws MissingTokenException
      * @throws SessionCreationException
      */
-    private function createRequestParams(FacetInterface $facets, $pageNumber, $limit)
+    public function createRequestParams(FacetInterface $facets, $pageNumber, $limit)
     {
         // Current store id value is unavailable
         $store = $this->nostoHelperScope->getStore();
@@ -261,68 +170,6 @@ class StateAwareCategoryService
             $previewMode,
             $token
         );
-    }
-
-
-    /**
-     * @param MerchandiseRequestParams $requestParams
-     * @return CategoryMerchandisingResult
-     * @throws AbstractHttpException
-     * @throws HttpResponseException
-     * @throws NostoException
-     */
-    private function getMerchandiseResults(MerchandiseRequestParams $requestParams)
-    {
-//        HttpRequest::buildUserAgent(
-//            'Magento',
-//            $this->nostoHelperData->getPlatformVersion(),
-//            "CMP_" . $this->cmHelperData->getModuleVersion()
-//        );
-
-        $categoryMerchandising = new BatchedCategoryMerchandising(
-            $requestParams->getNostoAccount(),
-            $requestParams->getCustomerId(),
-            $requestParams->getCategory(),
-            $requestParams->getPageNumber(),
-            $requestParams->getFacets()->getIncludeFilters(),
-            $requestParams->getFacets()->getExcludeFilters(),
-            '',
-            AbstractGraphQLOperation::IDENTIFIER_BY_CID,
-            $requestParams->isPreviewMode(),
-            $requestParams->getLimit(),
-            $requestParams->getBatchToken()
-        );
-
-        $result = $categoryMerchandising->execute();
-
-        $this->eventManager->dispatch(
-            PostRequestAction::DISPATCH_EVENT_NAME_POST_RESULTS,
-            [
-                PostRequestAction::DISPATCH_EVENT_KEY_RESULT => $result,
-                PostRequestAction::DISPATCH_EVENT_KEY_LIMIT => $requestParams->getLimit(),
-                PostRequestAction::DISPATCH_EVENT_KEY_PAGE => $requestParams->getPageNumber(),
-            ]
-        );
-
-        $this->trace(
-            'Got %d / %d (total) product ids from Nosto CMP for category "%s", using page num: %d, using limit: %d',
-            [
-                $result->getResultSet()->count(),
-                $result->getTotalPrimaryCount(),
-                $requestParams->getCategory(),
-                $requestParams->getPageNumber(),
-                $requestParams->getLimit()
-            ]
-        );
-        return $result;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getLastResult(): ?CategoryMerchandisingResult
-    {
-        return $this->lastResult;
     }
 
     /**
