+++ Model/Service/VisitSession/SessionService.php
@@ -1,5 +1,4 @@
-<?php /** @noinspection PhpUndefinedMethodInspection */
-
+<?php
 /**
  * Copyright (c) 2020, Nosto Solutions Ltd
  * All rights reserved.
@@ -35,57 +34,31 @@
  *
  */
 
-namespace Nosto\Cmp\Model\Service\Recommendation;
+namespace Nosto\Cmp\Model\Service\VisitSession;
 
-use Magento\Framework\Session\SessionManagerInterface;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Store\Model\Store;
+use Nosto\Cmp\Exception\SessionCreationException;
+use Nosto\NostoException;
+use Nosto\Operation\Session\NewSession;
+use Nosto\Types\Signup\AccountInterface;
 
 class SessionService
 {
-    /** @var SessionManagerInterface */
-    private $session;
-
     /**
-     * CmpSession constructor.
-     * @param SessionManagerInterface $session
+     * @param Store $store
+     * @param AccountInterface $nostoAccount
+     * @return mixed|null
+     * @throws SessionCreationException
      */
-    public function __construct(SessionManagerInterface $session)
+    public function getNewNostoSession(Store $store, AccountInterface $nostoAccount)
     {
-        $this->session = $session;
-    }
-
-    /**
-     * @param BatchModel $model
-     */
-    public function setBatchModel(BatchModel $model)
-    {
-        $this->session->start();
-        $this->session->setNostoCmpBatchSession($model); //@phan-suppress-current-line PhanUndeclaredMethod
-    }
-
-    /**
-     * @return BatchModel
-     */
-    public function getBatchModel()
-    {
-        $this->session->start();
-        return $this->session->getNostoCmpBatchSession(); //@phan-suppress-current-line PhanUndeclaredMethod
-    }
-
-    /**
-     * @param GraphQlParamModel $model
-     */
-    public function setGraphqlModel(GraphQlParamModel $model)
-    {
-        $this->session->start();
-        $this->session->setNostoCmpGraphqlSession($model); //@phan-suppress-current-line PhanUndeclaredMethod
-    }
-
-    /**
-     * @return GraphQlParamModel
-     */
-    public function getGraphqlModel()
-    {
-        $this->session->start();
-        return $this->session->getNostoCmpGraphqlSession(); //@phan-suppress-current-line PhanUndeclaredMethod
+        try {
+            $url = $store->getCurrentUrl();
+            $newSession = new NewSession($nostoAccount, $url, true);
+            return $newSession->execute();
+        } catch (NoSuchEntityException | NostoException $e) {
+            throw new SessionCreationException($store, $e);
+        }
     }
 }
