+++ Plugin/Api/Search/SearchResultSorter.php
@@ -38,23 +38,20 @@ namespace Nosto\Cmp\Plugin\Api\Search;
 
 use Magento\Framework\Api\Search\Document;
 use Magento\Framework\Api\Search\SearchResult;
-use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryService;
+use Nosto\Cmp\Model\Service\Merchandise\LastResult;
 
 class SearchResultSorter
 {
-    /**
-     * @var StateAwareCategoryService
-     */
-    private $categoryService;
+    /** @var LastResult  */
+    private $lastResult;
 
     /**
-     * SearchResultSorter constructor.
-     * @param StateAwareCategoryService $categoryService
+     * @param LastResult $lastResult
      */
     public function __construct(
-        StateAwareCategoryService $categoryService
+        LastResult $lastResult
     ) {
-        $this->categoryService = $categoryService;
+        $this->lastResult = $lastResult;
     }
 
     /**
@@ -100,7 +97,7 @@ class SearchResultSorter
      */
     private function getCmpSort()
     {
-        $categoryMerchandisingResult = $this->categoryService->getLastResult();
+        $categoryMerchandisingResult = $this->lastResult->getLastResult();
         if ($categoryMerchandisingResult !== null) {
             return $categoryMerchandisingResult->parseProductIds();
         }
@@ -113,7 +110,7 @@ class SearchResultSorter
      */
     private function getTotalPrimaryCount()
     {
-        $categoryMerchandisingResult = $this->categoryService->getLastResult();
+        $categoryMerchandisingResult = $this->lastResult->getLastResult();
         if ($categoryMerchandisingResult !== null) {
             return $categoryMerchandisingResult->getTotalPrimaryCount();
         }
