+++ Plugin/Framework/Search/Request/AbstractHandler.php
@@ -39,14 +39,17 @@ namespace Nosto\Cmp\Plugin\Framework\Search\Request;
 use Exception;
 use Magento\Store\Model\Store;
 use Nosto\Cmp\Exception\CmpException;
+use Nosto\Cmp\Exception\MissingAccountException;
+use Nosto\Cmp\Exception\MissingTokenException;
+use Nosto\Cmp\Exception\SessionCreationException;
 use Nosto\Cmp\Helper\Data as CmpHelperData;
 use Nosto\Cmp\Helper\SearchEngine;
 use Nosto\Cmp\Model\Facet\FacetInterface;
-use Nosto\Cmp\Model\Service\Recommendation\StateAwareCategoryService;
+use Nosto\Cmp\Model\Service\Merchandise\MerchandiseServiceInterface;
+use Nosto\Cmp\Model\Service\Merchandise\RequestParamsService;
 use Nosto\Cmp\Utils\Request as RequestUtils;
 use Nosto\Cmp\Utils\Search;
 use Nosto\Cmp\Utils\Traits\LoggerTrait;
-use Nosto\NostoException;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
 use Nosto\Tagging\Helper\Scope as NostoHelperScope;
 use Nosto\Tagging\Logger\Logger;
@@ -88,26 +91,32 @@ abstract class AbstractHandler
     protected $nostoHelperScope;
 
     /**
-     * @var StateAwareCategoryService
+     * @var MerchandiseServiceInterface
      */
-    protected $categoryService;
+    protected $merchandiseService;
 
     /**
-     * AbstractHandler constructor.
+     * @var RequestParamsService
+     */
+    private $requestParamService;
+
+    /**
      * @param SearchEngine $searchEngineHelper
      * @param NostoHelperAccount $nostoHelperAccount
      * @param NostoHelperScope $nostoHelperScope
      * @param CmpHelperData $cmpHelperData
-     * @param StateAwareCategoryService $categoryService
+     * @param MerchandiseServiceInterface $merchandiseService
+     * @param RequestParamsService $requestParamsService
      * @param Logger $logger
      */
     public function __construct(
-        SearchEngine $searchEngineHelper,
-        NostoHelperAccount $nostoHelperAccount,
-        NostoHelperScope $nostoHelperScope,
-        CmpHelperData $cmpHelperData,
-        StateAwareCategoryService $categoryService,
-        Logger $logger
+        SearchEngine               $searchEngineHelper,
+        NostoHelperAccount         $nostoHelperAccount,
+        NostoHelperScope           $nostoHelperScope,
+        CmpHelperData              $cmpHelperData,
+        MerchandiseServiceInterface $merchandiseService,
+        RequestParamsService       $requestParamsService,
+        Logger                     $logger
     ) {
         $this->loggerTraitConstruct(
             $logger
@@ -116,7 +125,8 @@ abstract class AbstractHandler
         $this->accountHelper = $nostoHelperAccount;
         $this->nostoHelperScope = $nostoHelperScope;
         $this->cmpHelperData = $cmpHelperData;
-        $this->categoryService = $categoryService;
+        $this->merchandiseService = $merchandiseService;
+        $this->requestParamService = $requestParamsService;
     }
 
     /**
@@ -282,15 +292,14 @@ abstract class AbstractHandler
      * @param $pageNum
      * @param $limit
      * @return array|null
-     * @throws NostoException
+     * @throws MissingAccountException
+     * @throws MissingTokenException
+     * @throws SessionCreationException
      */
     private function getCmpProductIds(FacetInterface $facet, $pageNum, $limit)
     {
-        $res = $this->categoryService->getPersonalisationResult(
-            $facet,
-            $pageNum,
-            $limit
-        );
+        $requestParams = $this->requestParamService->createRequestParams($facet, $pageNum, $limit);
+        $res = $this->merchandiseService->getMerchandiseResults($requestParams);
         return $res ? $res->parseProductIds() : null;
     }
 }
